+++ previous-versions/common/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v0.35.0] previous-versions/common
 ### Releases
 - [PyPI v0.35.0] https://pypi.org/project/codat-common/0.35.0 - previous-versions/common
+
+## 2023-10-01 00:19:15
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.92.2 (2.142.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.36.0] previous-versions/common
+### Releases
+- [PyPI v0.36.0] https://pypi.org/project/codat-common/0.36.0 - previous-versions/common